Java 6 / 64-bit work - Paul Loy nailed another pesky intValue.

+++ trunk/rococoa/src/main/java/org/rococoa/Rococoa.java
@@ -189,7 +189,7 @@ public abstract class Rococoa  {
         }
         
         public ProxyID(ID anotherID, OCInvocationCallbacks callbacks) {
-            super(anotherID.intValue());
+            super(anotherID);
             this.callbacks = callbacks;
         }
     }
